Table refactors for ease of use
